Fixed bug in MassAction_Sim (epidemics too small)
Recovery times were being recast from doubles to ints, which yielded smaller
than expected epidemics.  If the gamma (recovery) parameter was small, this
had a small effect; if large, it had a large effect.  For gamma = 1/6, beta
= 1/4, epidemics tended to be 1-5% too small.

+++ src/MassAction_Sim.h
@@ -48,7 +48,7 @@ class MassAction_Sim {
                                  // event queue
         priority_queue<Event, vector<Event>, compTime > EventQ;
         vector<int> States;      // list of states; 0 is "never been infected"
-        vector<int> Rec;         // list of recovery times
+        vector<double> Rec;      // list of recovery times
         double Now;
         int epi_size;            //size of epidemic
